🚚(website) move App related files to features/App
- rename SentryLoader to AppConfig
- move App related files to features/App
- move routes/AppRoutes to features/App
