--- src/frontend/apps/standalone_site/src/components/Sentry/index.tsx
@@ -1,1 +1,0 @@
-export * from './SentryLoader';
