+++ src/frontend/apps/standalone_site/src/features/App/App.spec.tsx
