+++ src/frontend/apps/standalone_site/src/features/App/index.ts
@@ -1,0 +1,1 @@
+export { default as App } from './App';
