+++ src/frontend/apps/standalone_site/src/index.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './index.css';
-import App from './App';
+import { App } from './features/App';
 
 ReactDOM.render(
   <React.StrictMode>
