+++ src/frontend/apps/standalone_site/src/routes/index.ts
@@ -1,3 +1,2 @@
-export { default as AppRoutes } from './AppRoutes';
 export * from './routes';
 export * from './useRoutes';
